Add files via upload
This three files are updated version of the project.  They been attached after the deadline, but hopefully they will be taken in consideration.
This version is not yet completed, but they  closer to fulfill the requirement. 
There are  missing some minor parts like moving the Issues and some update action. 
Intend to do so in the near future.